+++ packages/wme-ui/src/components/index.ts
@@ -55,8 +55,8 @@ export * from './chip';
 export { default as Pill } from './pill';
 export * from './pill';
 
-export { default as PillGroup } from './pillGroup';
-export * from './pillGroup';
+export { default as PillGroup } from './pill-group';
+export * from './pill-group';
 
 export { default as ProgressBar } from './progress-bar';
 export * from './progress-bar';
