+++ packages/wme-ui/src/components/pill-group/index.ts
@@ -1,0 +1,2 @@
+export * from './pill-group';
+export { default } from './pill-group';
