--- packages/wme-ui/src/components/pillGroup/index.ts
@@ -1,2 +1,0 @@
-export * from './pillGroup';
-export { default } from './pillGroup';
